+++ metabocore_app/config/urls.py
@@ -4,5 +4,5 @@ from django.views.generic import RedirectView
 
 urlpatterns = [
     path("", RedirectView.as_view(pattern_name="form_viewer:form_list", permanent=False)),
-    path("formatos/", include("form_viewer.urls")),
+    path("", include("form_viewer.urls")),
 ]